Enda mer koderydding av unødvendig kode
