--- src/main/kotlin/no/nav/personbruker/dittnav/brukernotifikasjonbestiller/common/exception/UnvalidatableRecordException.kt
@@ -1,5 +1,0 @@
-package no.nav.personbruker.dittnav.brukernotifikasjonbestiller.common.exception
-
-class UnvalidatableRecordException(message: String, cause: Throwable?) : AbstractPersonbrukerException(message, cause) {
-    constructor(message: String) : this(message, null)
-}
