+++ src/test/kotlin/no/nav/personbruker/dittnav/brukernotifikasjonbestiller/common/database/DatabaseTest.kt
@@ -65,5 +65,4 @@ class DatabaseTest {
             }
         }
     }
-
 }
