+++ src/test/kotlin/no/nav/personbruker/dittnav/brukernotifikasjonbestiller/done/DoneTestData.kt
@@ -1,0 +1,12 @@
+package no.nav.personbruker.dittnav.brukernotifikasjonbestiller.done
+
+import no.nav.brukernotifikasjon.schemas.input.DoneInput
+import java.time.Instant
+
+object DoneTestData {
+
+    fun doneInput(tidspunkt: Long = Instant.now().toEpochMilli()) =
+        DoneInput(
+            tidspunkt
+        )
+}
